requirements.txt
fastapi 
uvicorn[standard]
google-generativeai
python-dotenv
tldextract
httpx
youtube-transcript-api